% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atan3.R
\name{atan3}
\alias{atan3}
\title{tan inverse of ratio of two numbers, mapping from 0 to 2pi.}
\usage{
atan3(beta2, beta1)
}
\arguments{
\item{beta2}{The numerator for the tan inverse function}

\item{beta1}{The denominator for the tan inverse function}
}
\description{
The function computes the tan inverse of ratio of two numbers and maps it to the cycle
             from 0 to 2pi. A wrapper function for the atan function in R
}
\examples{
 atan3(4,1)
 atan3(2,3)

}
\author{
Kushal K Dey
}

