% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_ordering_full.R
\name{cell_ordering_full}
\alias{cell_ordering_full}
\title{Cell continuous ordering - post processing of cell time class ordering}
\usage{
cell_ordering_full(signal_intensity, num_genes)
}
\arguments{
\item{num_genes}{The number of genes in the data.}

\item{signal_intensity:}{The signal intensity for the cells across time classes (N x T) matrix,
where T represents the number of time classes (equal to celltime_levels)}
}
\value{
Returns the full cell relative ordering of the cells on the
         cell cycle.
}
\description{
The function takes the output from the cell_phase_ordering function and provides a
             more continuous ordering of the cells based on the signal intensity of the cells in the
             different cell time classes as in the cell_phase_ordering function.
}
\examples{

}
\author{
Kushal K Dey
}
