% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/np_loglik_cellcycle.R
\name{np_loglik_cell_cycle}
\alias{np_loglik_cell_cycle}
\title{Loglikelihood of of sinusoidal genes in cell cycle}
\usage{
np_loglik_cell_cycle(cycle_data, cell_times, fit, sigma)
}
\arguments{
\item{cell_times}{: a N x 1 vector of cell times}

\item{cycle_data:}{a N x G matrix, where N is number of cells, G number of genes}

\item{fit:}{a celltime_levels(C) x G matrix of model fit from nonparametric cellcycleR}

\item{sigma:}{the G x 1 vector of gene variation}
}
\description{
Computes the loglikelihood of all the cells in the cycle under nonparametric smoothing

 @author  Kushal K Dey

 @export

 @examples
 G <- 500;
 num_cells <- 400;
 amp_genes <- rep(10, G);
 phi_genes <- runif(G, 0, 2*pi)
 sigma_genes <- rchisq(G, 4);
 cell_times_sim <- sample(seq(0,2*pi, 2*pi/(num_cells-1)), num_cells, replace=FALSE);
 cycle_data <- sim_sinusoidal_cycle(G, amp_genes, phi_genes, sigma_genes, cell_times_sim);

 loglik_cell_cycle_np(cycle_data, cell_times, amp_genes, phi_genes, sigma_genes)
}

