% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/np_loglik_cellcycle.R
\name{np_loglik_cellcycle}
\alias{np_loglik_cellcycle}
\title{Loglikelihood of of sinusoidal genes in cell cycle}
\usage{
np_loglik_cellcycle(cycle_data, cell_times, fit, sigma)
}
\arguments{
\item{cell_times}{: a N x 1 vector of cell times}

\item{cycle_data:}{a N x G matrix, where N is number of cells, G number of genes}

\item{fit:}{a celltime_levels(C) x G matrix of model fit from nonparametric cellcycleR}

\item{sigma:}{the G x 1 vector of gene variation}
}
\description{
Computes the loglikelihood of all the cells in the cycle under nonparametric smoothing

 @author  Kushal K Dey

 @export
}

