% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_sinusoidal_cycle.R
\name{sim_sinusoidal_cycle}
\alias{sim_sinusoidal_cycle}
\title{Simulate sinusoidal gene expression along cell phases on cell cycle}
\usage{
sim_sinusoidal_cycle(num_genes, amp_genes, phi_genes, sigma_genes, cell_times,
  correlation = NULL)
}
\arguments{
\item{num_genes}{The number of sinusoidal genes to simulate}

\item{amp_genes}{The amplitude vector of the sinusoidal genes (of length equal to num_genes)}

\item{phi_genes}{The phase angle vector of the sinusoidal genes (of length equal to num_genes)}

\item{sigma_genes}{The noise variation vector of the sinusoidal genes (of length equal to num_genes)}

\item{cell_times}{The phases of the cells on the cell cycle (a vector of values between 0 to 2 pi degrees)
The length of cell_times corresponds to number of single cells considered for simulation (say N).}
}
\value{
A matrix of size N x num_genes.
}
\description{
The function simulates sinusoidal gene patterns over a cell cycle
			  	with the cell phases provided by the user.
}
\examples{

G <- 500
num_cells <- 400
amp_genes <- rep(10, G)
phi_genes <- runif(G, 0, 2*pi)
sigma_genes <- rchisq(G, 4)
cell_times_sim <- sample(seq(0,2*pi, 2*pi/(num_cells-1)), num_cells, replace=FALSE);

# Assume uncorrelated gene expression
cycle_data <- sim_sinusoidal_cycle(G, amp_genes, phi_genes,
                                   sigma_genes, cell_times_sim,
                                   correlation = NULL)

# Assume correlation gene expression
cycle_data <- sim_sinusoidal_cycle(G, amp_genes, phi_genes,
                                   sigma_genes, cell_times_sim,
                                   correlation = .3)
}
\author{
Kushal K Dey, Joyce Hsiao
}

