% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sin_loglik_cellcycle.R
\name{sin_loglik_cellcycle}
\alias{sin_loglik_cellcycle}
\title{Loglikelihood of sinusoidal model}
\usage{
sin_loglik_cellcycle(cycle_data, cell_times, amp, phi, sigma, freq = 1)
}
\arguments{
\item{cycle_data}{a N x G matrix, where N is number of cells, G number of genes}

\item{cell_times}{a N x 1 vector of cell times}

\item{amp}{the amplitude vector (G x 1) over the genes}

\item{phi}{the G x 1 vector of phase values over genes}

\item{sigma:}{the G x 1 vector of gene variation}

\item{freq:}{frequency of the signal of the gene. Defaults to 1.}
}
\description{
Computes the loglikelihood of all the cells in the cycle under sinusoidal
             gene expression patterns.

 @author  Kushal K Dey

 @examples
 G <- 500;
 num_cells <- 400;
 amp_genes <- rep(10, G);
 phi_genes <- runif(G, 0, 2*pi)
 sigma_genes <- rchisq(G, 4);
 cell_times_sim <- sample(seq(0,2*pi, 2*pi/(num_cells-1)), num_cells, replace=FALSE);
 cycle_data <- sim_sinusoidal_cycle(G, amp_genes, phi_genes, sigma_genes, cell_times_sim);

 sing_loglik_cellcycle(cycle_data, cell_times, amp_genes, phi_genes, sigma_genes)

 @export
}
