% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/np_cell_ordering_class.R
\name{np_cell_ordering_class}
\alias{np_cell_ordering_class}
\title{Nonparametric cell ordering into different phases on the cell cycle}
\usage{
np_cell_ordering_class(cycle_data, celltime_levels, num_iter,
  method = c("LOESS", "B-spline", "Wavelet"), save_path = NULL)
}
\arguments{
\item{num_iter}{The number of iterations to run for the Gibbs sampler scheme}

\item{save_path}{The file path to save the RDA file containing the information on gene characteristics
and the cell ordering times. Default is NULL in which case, it will not save output}

\item{cycle_data:}{a N x G matrix, where N is number of cells, G number of genes}

\item{celltime_levels:}{The number of phase classes chosen (The deafult is 100). It splits up
circle or the range (0, 2pi) into celltime_levels uniform splits.}
}
\value{
Returns a list containing the following items
 \item{cell_times}{estimated cell times from Gibbs sampler}
 \item{sigma}{The estimated non-signal or noise variation of the genes}
 \item{loglik}{The model log likelihood of the fit}
 \item{signal_intensity}{The intensity of each cell in each of the cell times classes}
}
\description{
The function runs a Gibbs sampler for a number of iterations to obtain the estimates
            of cell times on cell cycle (modulo rotation) as well as the gene features (smoothing fit)
}
\author{
Kushal K Dey
}
