% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sin_cell_ordering_class.R
\name{sin_cell_ordering_class}
\alias{sin_cell_ordering_class}
\title{Sinusoidal model fitting}
\usage{
<<<<<<< HEAD
sin_cell_ordering_class(cycle_data, celltime_levels, num_iter = NULL,
  save_path = NULL, fix.phase = FALSE, phase_in = NULL, verbose = FALSE,
  tol = 0.01, maxiter = 500)
=======
sin_cell_ordering_class(cycle_data, celltime_levels, num_iter,
  save_path = NULL, fix.phase = FALSE, phase_in = NULL, verbose = FALSE)
>>>>>>> upstream/master
}
\arguments{
\item{cycle_data}{a N x G matrix, where N is number of cells, G number of genes}

\item{celltime_levels}{The number of phase classes chosen (The deafult is 100). It splits up
circle or the range (0, 2pi) into celltime_levels uniform splits.}

\item{num_iter}{Number of iterations. If not specified, number of iteration
is set to be maxiter (the maximum number of iteratoins).}

\item{save_path}{The file path to save the RDA file containing the information on gene characteristics
and the cell ordering times. Default is NULL in which case, it will not save output}

\item{fix.phase}{if TRUE, the phase will be fixed in inference for the genes, default is FALSE}

\item{phase_in}{if fix.phase is TRUE, then phase_in is G x 1 vector of user input gene phases.
Default is NULL as is the case if fix.phase=FALSE.}
<<<<<<< HEAD
=======

\item{verbose}{if TRUE, prints the loglikelihood at each step/iteration. If FALSE,just prints final loglikelihood.}

\item{cycle_data:}{a N x G matrix, where N is number of cells, G number of genes}

\item{celltime_levels:}{The number of phase classes chosen (The deafult is 100). It splits up
circle or the range (0, 2pi) into celltime_levels uniform splits.}
>>>>>>> upstream/master
}
\value{
Returns a list containing the following items
 \item{cell_times}{estimated cell times from Gibbs sampler}
 \item{amp}{The estimated amplitudes of the genes}
 \item{phi}{The estimated phase angles of the genes}
 \item{sigma}{The estimated non sinusoid signal variation of the genes}
 \item{loglik}{The model log likelihood of the fit}
 \item{signal_intensity}{The intensity of each cell in each of the cell times classes}
}
\description{
This function fits a sinsoidal model to gene expression data and 
             computes cell-order estimates on cell-cycle (modulo rotation),
             as well as the gene-specific cell-cycle features, including amplitude,
             phase and noise variation.
}
\author{
Kushal K Dey, Joyce Hsiao
}

